Added interfaces for events
