+++ pow/src/main/java/tech/pegasys/artemis/pow/event/ValidatorRegistration.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2018 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.artemis.pow.event;
+
+import tech.pegasys.artemis.pow.api.ValidatorRegistrationEvent;
+import tech.pegasys.artemis.pow.contract.ValidatorRegistrationContract;
+
+public class ValidatorRegistration implements ValidatorRegistrationEvent {
+
+  private ValidatorRegistrationContract.Eth1DepositEventResponse response = null;
+
+  public ValidatorRegistration(
+      ValidatorRegistrationContract.Eth1DepositEventResponse response) {
+    this.response = response;
+  }
+
+  public ValidatorRegistrationContract.Eth1DepositEventResponse getResponse() {
+    return response;
+  }
+}
